requirements.txt
Pillow==7.1.2
astroid==2.4.1
attrs==19.3.0
coverage==5.1
cycler==0.10.0
future==0.18.2
importlib-metadata==1.6.1
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
more-itertools==8.3.0
numpy==1.18.5
packaging==20.4
pluggy==0.13.1
py==1.8.1
pylint==2.5.2
pyparsing==2.4.7
pytest-cov==2.9.0
pytest==5.4.3
python-dateutil==2.8.1
scipy==1.4.1
six==1.15.0
toml==0.10.1
torch==1.5.0
torchvision==0.6.0
typed-ast==1.4.1
wcwidth==0.2.3
wrapt==1.12.1
zipp==3.1.0
